+++ src/components/VideoPlayer/videoPlayerChangelogData.ts
@@ -10,6 +10,15 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "2025-06-12",
+    version: "3.6.3",
+    type: "Update",
+    affects: ["Accessibility"],
+    notes: [
+      "Display a fallback warning message on JavaScript-disabled browsers.",
+    ],
+  },
+  {
     date: "2024-03-14",
     version: "3.0.0",
     type: "Update",